split generation, intersection, shading, post processing

+++ src/cpp/util/util.cpp
@@ -7,7 +7,7 @@ std::vector<char> readFile(const std::string& filename) {
         throw std::runtime_error("failed to open file!");
     }
 
-    size_t fileSize = (size_t) file.tellg();
+    size_t fileSize = file.tellg();
     std::vector<char> buffer(fileSize);
 
     file.seekg(0);
